% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_integral.R
\name{calcular_integral}
\alias{calcular_integral}
\title{Calcular integral definida en el [0,1]}
\usage{
calcular_integral(g, n = 1000)
}
\arguments{
\item{g}{funcion definida en el [0,1]}

\item{n}{numero de uniformes generadas}
}
\value{
una lista con el valor estimado de la integral y su error asociado.
}
\description{
Esta funcion utiliza un metodo de simulacion
por Monte Carlo para calcular la integral definida
en el rango [0,1] de una funcion \code{g}
generando \code{n} variables aleatorias uniformes.
}
\examples{
calcular_integral(g = \(x) exp(x))

}
