% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_discreta_ar.R
\name{generar_discreta_ar}
\alias{generar_discreta_ar}
\title{Generar variable aleatoria discreta}
\usage{
generar_discreta_ar(x, px)
}
\arguments{
\item{x}{Valores posibles de X.}

\item{px}{Función de probabilidad p(x) en el mismo orden que \code{x}.}
}
\value{
Número aleatorio de la distribución especificada.
}
\description{
Esta función genera una variable aleatoria discreta X
tabulada usando el método de aceptación/rechazo.
}
\details{
Esta función no requiere especificar la ley
de probabilidad q(y) de la variable auxiliar Y porque
siempre utiliza la distribución uniforme discreta
para generar posibles valores de X.
}
\examples{
generar_discreta_ar(x = 1:4, px = c(0.50,0.30,0.15,0.05))

}
