% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_discreta_ti.R
\name{generar_discreta_ti}
\alias{generar_discreta_ti}
\title{Generar variable aleatoria discreta}
\usage{
generar_discreta_ti(x, px)
}
\arguments{
\item{x}{Valores posibles de X.}

\item{px}{Función de probabilidad p(x) en el mismo orden que \code{x}.}
}
\value{
Número aleatorio de la distribución especificada.
}
\description{
Esta función genera una variable aleatoria discreta X
tabulada usando el método de la transformada inversa.
}
\examples{
generar_discreta_ti(x = 1:4, px = c(0.50,0.30,0.15,0.05))

}
