% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_uniforme_discreta.R
\name{generar_uniforme_discreta}
\alias{generar_uniforme_discreta}
\title{Generar variable aleatoria uniforme discreta}
\usage{
generar_uniforme_discreta(a, b)
}
\arguments{
\item{a}{Límite inferior.}

\item{b}{Límite superior.}
}
\value{
Número aleatorio entero entre \code{a} y \code{b}.
}
\description{
Esta función genera una variable aleatoria
uniforme discreta con igual probabilidad
en el intervalo [a,b].
}
\examples{
# tirar un dado legal:
generar_uniforme_discreta(1,6)

}
