% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_integral.R
\name{calcular_integral}
\alias{calcular_integral}
\title{Calcular integral definida en el [0,1] por Monte Carlo}
\usage{
calcular_integral(g, n = 1000)
}
\arguments{
\item{g}{Funcion definida en el [0,1].}

\item{n}{Numero de simulaciones.}
}
\value{
Una lista con el valor \code{estimado} de la integral y su \code{error} asociado.
}
\description{
Esta funcion calcula la integral definida
en el rango [0,1] de una funcion \code{g}
simulando \code{n} variables aleatorias U(0,1).
}
\examples{
calcular_integral(g = \(x) exp(x))

}
