% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_numeros_aleatorios.R
\name{generar_numeros_aleatorios}
\alias{generar_numeros_aleatorios}
\title{Generar numeros aleatorios}
\usage{
generar_numeros_aleatorios(n, x0, a, c, m)
}
\arguments{
\item{n}{largo de la secuencia}

\item{x0}{valor inicial o semilla}

\item{a}{multiplicador}

\item{c}{constante aditiva}

\item{m}{modulo}
}
\value{
Secuencia de numeros aleatorios en el intervalo [0,1] de largo \code{n}.
}
\description{
Esta funcion genera una secuencia rectangular de numeros aleatorios
a partir de un generador congruencial lineal:
\deqn{x_n = (a \cdot x_{n-1} + c) \text{ mod } m}
}
\examples{
generar_numeros_aleatorios(n = 10, x0 = 5, a = 3, c = 0, m = 150)

}
